Move date_time_handling to utils
